added pagination and laravel ACL

+++ app/Http/Controllers/API/UserController.php
@@ -22,7 +22,7 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        return User::latest()->paginate(5);
     }
 
     /**
@@ -51,6 +51,40 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    public function profile()
+    {
+        return auth('api')->user();
+    }
+
+    public function updateProfile(Request $request)
+    {
+        $user = auth('api')->user();
+        $this->validate($request, [
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,'.$user->id,
+            'password' => 'sometimes|required|min:6'
+        ]);
+        $currentPhoto = $user->photo;
+        if($request->photo != $currentPhoto){
+            if($request->photo){
+                $name = time().'.'.explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
+            }
+            \Image::make($request->photo)->save(public_path('images/profile/').$name);
+            $request->merge(['photo'=>$name]);
+
+            $userPhoto = public_path('images/profile/').$currentPhoto;
+            if(file_exists($userPhoto)){
+                @unlink($userPhoto);
+            }
+        }
+        if(!empty($request->password)){
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+        $user->update($request->all());
+        return ['message'=>'Success'];
+
+    }
+
     public function show($id)
     {
         //
@@ -84,10 +118,24 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
 
         $user->delete();
 
         return ['deleted' => 'User Successfully Deleted'];
+    }
+
+    public function search(Request $request) {
+        if($search = $request->get('q')){
+            $users = User::where(function($query) use ($search){
+                $query->where('name', 'LIKE', "%$search%")
+                    ->orWhere('email', 'LIKE', "%$search%")
+                    ->orWhere('type', 'LIKE', "%$search%");
+            })->paginate(5);
+        }else{
+            $users = User::latest()->paginate(5);
+        }
+        return $users;
     }
 }
